requirements.txt
streamlit>=1.28.0
openai>=1.0.0
stripe>=6.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
